Move signature/selector computation functions to abi-utils
