+++ packages/abi-utils/lib/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
 export * from "./normalize";
+export * from "./signature";
 import * as Arbitrary from "./arbitrary";
 export { Arbitrary };
