+++ packages/abi-utils/package.json
@@ -26,7 +26,8 @@
   "dependencies": {
     "change-case": "3.0.2",
     "faker": "5.5.3",
-    "fast-check": "3.1.1"
+    "fast-check": "3.1.1",
+    "web3-utils": "1.7.4"
   },
   "devDependencies": {
     "@fast-check/jest": "^1.0.1",
