+++ packages/codec/lib/evm/utils.ts
@@ -4,10 +4,11 @@ const debug = debugModule("codec:evm:utils");
 import BN from "bn.js";
 import Web3Utils from "web3-utils";
 import * as Conversion from "@truffle/codec/conversion";
+import { ShortSelectorSize } from "@truffle/abi-utils";
 
 export const WORD_SIZE = 0x20;
 export const ADDRESS_SIZE = 20;
-export const SELECTOR_SIZE = 4; //function selectors, not event selectors
+export const SELECTOR_SIZE = ShortSelectorSize; //equals 4; function & error selectors, not event selectors
 export const PC_SIZE = 4;
 export const MAX_WORD = new BN(-1).toTwos(WORD_SIZE * 8);
 export const ZERO_ADDRESS = "0x" + "00".repeat(ADDRESS_SIZE);
